Fix EBS attachment for additional volumes
Using hardcoded NVME device paths does not work because they can
change at boot time, or can be presented to the system in an
unexpected order.

To fix this, we pass the EBS device path (/dev/sdX) which is
explicitly assigned by terraform and use the nvme-cli tool for
finding out the right NVME device, based on EBS device path.

Once we have the right NVME device, we can find out disk UUID and
use it for updating /etc/fstab.