`inventory.yml.tftpl` default template for edb-ansible and related templating changes (#70)
Templates can now be passed in through the CLI with `--user-templates`
or `ET_USER_TEMPLATES` as a list of template files or a directory with
template files. Users can still make use of the infrastructure file to
specify templates paths but they will be copied over to the projects
template directory and removed from the final terraform variables file
as it is unused within terraform.
All templates will be rendered if they are within the projects
`template` directory, if it ends with the `.tftpl` extension.
Preferably, templates should not cause an infrastructure configuration
to fail if is provided as a sample or a default.

`inventory.yml.tftpl` has been added as an example which works with
`edb-ansible`. To better work alongside ansible, additional outputs were
passed through with our machines so they could be added to the inventory
file, including `ansible_user` and `ansible_ssh_private_key_file`.

Fixes: AWS/Azure/Gcloud - `setup_volume.sh` script updated to handle dropping of `/dev/` from device name when checking with nvme cli tools.