remove spot max_price
Simplify the use of spot instances, such that "spot: true" can be used
to request a spot instance.  The purpose is to take advantage of spot
instances for short lived systems, but pay on-demand pricing if need be,
while lowering the risk of losing a system while testing.

Note it is still possible to lose a spot instance system even if pricing
reaches the on-demand level.