Refactor nvme device identification
The regex was too greedy for device names with 4 vs 3 characters.
Expand this into a for loop to explicitly compare exact device names.
(I suppose this still could have been a regex.)

Added an additional check to throw a non-zero return code if the nvme
device could not be identified.