Github generate action - By default, use the "cloud" option which maps to cloud providers backend storage offering. Backend mappings for "cloud": aws => s3 | azure => azurerm | gcloud => gcs If a specific backend is needed, the 'remote_state' can be set to any terraform backend. This requires setting up the backend config as a list of variables or a file to pass to the apply action for use with 'terraform init'.
Use of local is discouraged since it might require manual intervention if a workflow fails.
This would also require backing up of the state, which should be protected at all times since it may contain secrets and is needed to destroy resources.
Since all resources are tagged, this might be acceptable if some other tool is used to clean up resources.